bot_move/requirements.txt
python-telegram-bot==13.15
psycopg2-binary==2.9.6
requests==2.28.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0